move underscore to its old place and remove unnecessary packages
